fix return value for Day12 stubs

+++ app/src/main/groovy/Day12.groovy
@@ -1,5 +1,5 @@
 class Day12Groovy {
-    static Integer findAverage(List<Integer> list) {
+    static BigDecimal findAverage(List<Integer> list) {
         return null;
     }
 }
