+++ app/src/main/kotlin/Day12.kt
@@ -1,1 +1,3 @@
-fun findAverage(list : List<Int>) : Int? = null
+import java.math.BigDecimal
+
+fun findAverage(list : List<Int>) : BigDecimal? = null
